Create header and hero section
Header and hero section are combined, the hero-text, secondary hero text
and signup button are grouped into a parent container. While, the
placeholder is separated. Making it look like the placeholder and the
container .hero-signup-container are child flex of the .hero parent flex
container.

Add a new .big-container that encapsulate both .header and .hero
containers.