Restructure css and add new image
For now header and hero section is completed.
Like I've said in my previous commit, I've encapsulated the header and
hero section. I included a gap between them.

A padding is added in the .big-container.

Temporarily fixed some flex alignments between the placeholder and
hero-text.

Switched from space-evenly to space-between in justfify-content of
.logo-links container.

To be continued is the content container and call-to-action.