Fix container padding
Fix .big-container padding, reduced to padding: 10px 48px 48px 48px;

The reason is when I started working on it the screen resolution is set
to 1920x1080 which is larger than the common one seen in most browser
screens.

Switched to 1280x720 desktop screen size.

Soon, I'll add a responsiveness to it.

Edit README.md