Add call-to-action and footer section
Added call to action and footer section.

I've added an `<em>` tag though I do not know if should I remove the
`<blockquote>` or not.

Changed the title in the information section.

As for the footer, I didn't create another `<div>` for the text itself.
I figured out that there is no other flex item or element exists other
than the copyright text.